requirements.txt
pyautogen==0.9.9
python-dotenv==1.0.0
flask==2.3.3